requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.1

# Validation
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# HTTP client for Shopify
httpx==0.27.2
aiohttp==3.11.0

# Encryption
cryptography==43.0.1

# Redis (for queues - optional)
redis==5.2.0
hiredis==2.3.2

# Utilities
python-dateutil==2.9.0.post0
